==> 02_analysis/03_top_genes.ipynb <==
import pandas as pd
import seaborn as sns
import numpy as np
import scipy as sp
import matplotlib.pyplot as plt
from scipy import stats
import pingouin as pg
import glob, os
from statsmodels.sandbox.stats.multicomp import multipletests
import statsmodels.formula.api as sm
data='../data/'

%%time
dt=pd.read_csv(data+'results/top_cgs_mesa.csv',index_col=0)
dt=dt.rename(columns={'visit':'cohort'})
d1=dt[dt['cohort']==1].copy()
d2=dt[dt['cohort']==2].copy()
d1['age_c']=pd.qcut(d1['age'],3,labels=['young','','old']).astype(str)
d2['age_c']=pd.qcut(d2['age'],3,labels=['young','','old']).astype(str)
d1['cohort']='mesa1'
d2['cohort']='mesa2'
d3=pd.read_csv(data+'results/top_cgs_ppmi.csv',index_col=0)
d3['cohort']='ppmi'
d3['age_c']=pd.qcut(d3['age'],3,labels=['young','','old']).astype(str)
dd=pd.concat([d1,d2,d3])
dd=dd[dd['age_c']!=''].copy()
dd.to_csv(data+'results/top_cgs_mesa+ppmi.csv')
fig2c_top=dd.drop(['age','sid'],axis=1).reset_index(drop=True)
fig2c_top.to_csv(data+'figs/2c_top_part.csv')

v1=dt[dt['cohort']==1].copy()
print(len(v1),len(v2))
slope, intercept, r_value, p_value, std_err = stats.linregress(v1['age'],v1['cg22213242'])
v1['d']=v1['cg22213242']-slope*v1['age']-intercept
v1['dc']=pd.qcut(v1['d'],10,labels=range(10))
v1['AA']=v1['dc']==9
v2=dt[dt['cohort']==2].copy()
v2['AA']=False
v2.loc[v2['sid'].isin(v1[v1['AA']].sid),'AA']=True
v2['AA']=v2['AA'].replace({True:'Highest CD248\n DNAm in 2000',False: 'Others'})
fig2d1=v1[['age','cg22213242','AA']].reset_index(drop=True)
fig2d2=v2[['age','cg22213242','AA']].reset_index(drop=True)
fig2d1.to_csv(data+'figs/2d1.csv')
fig2d2.to_csv(data+'figs/2d2.csv')

dt=pd.read_csv(data+'results/top_rna.csv',index_col=0)
d1=dt[dt['cohort']=='MESA1'].copy()
d2=dt[dt['cohort']=='MESA2'].copy()
d1['age_c']=pd.qcut(d1['age'],3,labels=['young','','old']).astype(str)
d2['age_c']=pd.qcut(d2['age'],3,labels=['young','','old']).astype(str)
d1['cohort']='mesa1'
d2['cohort']='mesa2'
d3=dt[dt['cohort']=='PPMI'].copy()
d3['age_c']=pd.qcut(d3['age'],3,labels=['young','','old']).astype(str)
dd=pd.concat([d1,d2,d3])
dd=dd[dd['age_c']!=''].copy()
for gen in ['ELOVL2','KLF14','FHL2','CD248','LRRN3','NELL2','SATB1']:
    dd[gen]=np.log10(dd[gen]+1)
fig2c_bottom=dd.drop(['sid','age'],axis=1).reset_index(drop=True)
fig2c_bottom.to_csv(data+'figs/2c_bottom.csv')
